open drain led configuration code added

+++ drivers/inc/stm32f407xx_gpio_driver.h
@@ -103,7 +103,7 @@ void GPIO_ToggleOutputPin(GPIO_RegDef_t *pGPIOx, uint8_t PinNumber);
 /* IRQ configuration and ISR handling */
 
 void GPIO_IRQInterruptConfig(uint8_t IRQNumber, uint8_t EnorDi);
-void GPIO_IRQPriorityConfig (uint8_t IRQNumber, uint8_t IRQPriority);
+void GPIO_IRQPriorityConfig (uint8_t IRQNumber, uint32_t IRQPriority);
 void GPIO_IRQHandling(uint8_t PinNumber);
 
 #endif /* INC_STM32F407XX_GPIO_DRIVER_H_ */
